alamnet/pyproject.toml
[tool.poetry]
name = "alamnet"
version = "0.1.0"
description = ""
authors = ["alamantik18 <59261897+alamantik18@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
djangorestframework = "^3.14.0"
pillow = "^9.4.0"
psycopg2 = "^2.9.5"
djoser = "^2.1.0"
drf-yasg = "^1.21.4"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

